requirements-simple.txt
# Core Data Engineering
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2

# Database & Storage
duckdb==0.9.2

# API & Web
requests==2.31.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0

# Data Validation (versão compatível)
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3

# Date & Time
python-dateutil==2.8.2

# Jupyter (opcional - pode comentar se não usar)
# jupyter==1.0.0
# ipykernel==6.28.0
